requirements.txt
Flask
google-generativeai
Flask-Cors
gunicorn
Flask-SQLAlchemy
psycopg2-binary
gevent